upd gorm to v2

+++ storages/dbs/friendreldao.go
@@ -54,7 +54,7 @@ func (rel FriendRelDao) QueryFriendRels(appkey, userId string, startId, limit in
 	}
 	condition = condition + " and id>?"
 	params = append(params, startId)
-	err := dbcommons.GetDb().Where(condition, params...).Order("id asc").Limit(limit).Find(&items).Error
+	err := dbcommons.GetDb().Where(condition, params...).Order("id asc").Limit(int(limit)).Find(&items).Error
 	if err != nil {
 		return nil, err
 	}
@@ -82,7 +82,7 @@ func (rel FriendRelDao) QueryFriendRelsWithPage(appkey, userId string, orderTag 
 	}
 
 	var items []*FriendRelWithUser
-	err := dbcommons.GetDb().Raw(sql, params...).Order("case when u.pinyin REGEXP '^[A-Za-z]' then 1 else 0 end desc, u.pinyin asc").Offset((page - 1) * size).Limit(size).Find(&items).Error
+	err := dbcommons.GetDb().Raw(sql, params...).Order("case when u.pinyin REGEXP '^[A-Za-z]' then 1 else 0 end desc, u.pinyin asc").Offset(int((page - 1) * size)).Limit(int(size)).Find(&items).Error
 	if err != nil {
 		return nil, err
 	}
@@ -115,7 +115,7 @@ type FriendRelWithUser struct {
 func (rel FriendRelDao) SearchFriendsByName(appkey, userId string, nickname string, startId, limit int64) ([]*models.User, error) {
 	sql := fmt.Sprintf("select r.*,u.nickname,u.user_portrait,u.user_type,u.pinyin from %s as r left join %s as u on r.app_key=u.app_key and r.friend_id=u.user_id where r.app_key=? and r.user_id=? and r.id>? and (u.nickname like ? or r.display_name like ? )", rel.TableName(), UserDao{}.TableName())
 	var items []*FriendRelWithUser
-	err := dbcommons.GetDb().Raw(sql, appkey, userId, startId, "%"+nickname+"%", "%"+nickname+"%").Order("r.id asc").Limit(limit).Find(&items).Error
+	err := dbcommons.GetDb().Raw(sql, appkey, userId, startId, "%"+nickname+"%", "%"+nickname+"%").Order("r.id asc").Limit(int(limit)).Find(&items).Error
 	ret := []*models.User{}
 	if err == nil {
 		for _, item := range items {
